+++ R/sample_belief_space.R
@@ -1,9 +1,10 @@
 #' Sample from the Belief Space
 #'
-#' Sample randomly (uniform) or regularly spaced points from the projected
-#' belief space.
+#' Sample points from belief space using a several sampling strategies.
 #'
-#' Several sampling methods are implemented:
+#' The purpose of sampling from the belief space is to provide good coverage or to sample belief points
+#' that are more likely to be encountered (see trajectory method).
+#' The following sampling methods are available:
 #' 
 #' * `'random'` samples uniformly sample from the projected belief space using
 #' the method described by Luc Devroye (1986). 
@@ -14,16 +15,21 @@
 #' 
 #' * `'vertices'` only samples from the vertices of the belief space.
 #'
+#' * `"trajectories"` returns the belief states encountered in `n` trajectories of length `horizon` starting at the
+#' model's initial belief. Thus it returns `n` x `horizon` belief states and will contain duplicates. 
+#' Projection is not supported for trajectories. Additional
+#' arguments can include the simulation `horizon` and the start `belief` which are passed on to [simulate_POMDP()].
+#'
 #' @family POMDP
 #' 
 #' @param model a unsolved or solved [POMDP].
 #' @param projection Sample in a projected belief space. All states not
 #' included in the projection are held at a belief of 0. `NULL` means no
 #' projection.
-#' @param n size of the sample.
+#' @param n size of the sample. For trajectories, it is the number of trajectories.
 #' @param method character string specifying the sampling strategy. Available
-#' are `"random"`, `"regular"`, and `"vertices"`.
-#' @param ... further arguments are ignored.
+#' are `"random"`, `"regular"`, `"vertices"`, and `"trajectories"`.
+#' @param ... for the trajectory method, further arguments are passed on to [simulate_POMDP()]. Further arguments are ignored for the other methods. 
 #' @return Returns a matrix. Each row is a sample from the belief space.
 #' @author Michael Hahsler
 #' @references Luc Devroye, Non-Uniform Random Variate Generation, Springer
@@ -33,10 +39,13 @@
 #'
 #' sample_belief_space(Tiger, n = 5)
 #' sample_belief_space(Tiger, n = 5, method = "regular")
+#' sample_belief_space(Tiger, n = 5, horizon = 5, method = "trajectories") 
 #'
-#' # sample and calculate the reward for a solve POMDP
+#' # sample and calculate the reward for a solved POMDP
 #' sol <- solve_POMDP(Tiger)
-#' reward(sol, belief = sample_belief_space(sol, n = 5, method = "regular"))
+#' samp <- sample_belief_space(sol, n = 5, method = "regular")
+#' rew <- reward(sol, belief = samp)
+#' cbind(samp, rew)
 #' @export
 sample_belief_space <-
   function(model,
@@ -44,7 +53,10 @@ sample_belief_space <-
     n = 1000,
     method = "random", ...) {
     method <-
-      match.arg(method, choices = c("random", "regular", "vertices"))
+      match.arg(method, choices = c("random", "regular", "vertices", "trajectories"))
+    
+    # check 
+    if(method == "trajectories" && !is.null(projection)) stop("projection is not supported for method 'trajectories'.")
     
     if (is.null(projection))
       projection <- seq(length(model$states))
@@ -54,16 +66,21 @@ sample_belief_space <-
     if (d < 2)
       stop("Projection needs to be on 2 or more states.")
     
-    # empty belief states
-    belief_states <-
-      matrix(0, nrow = n, ncol = length(model$states))
-    colnames(belief_states) <- model$states
+    if (any(duplicated(projection)) || any(is.na(match(projection, seq_along(model$states)))))
+      stop("illegal specification for projection.")
     
-    switch(method,
+    # empty belief states
+
+    
+    belief_states <- switch(method,
       random = {
         # uniformly sample from a simplex.
         # https://cs.stackexchange.com/questions/3227/uniform-sampling-from-a-simplex)
         # Luc Devroye, Non-Uniform Random Variate Generation, Springer Verlag, 1986.
+        belief_states <-
+          matrix(0, nrow = n, ncol = length(model$states))
+        colnames(belief_states) <- model$states
+        
         m <-
           cbind(0, matrix(stats::runif(n * (d - 1)), ncol = d - 1), 1)
         belief_states[, projection] <-
@@ -73,9 +90,14 @@ sample_belief_space <-
             FUN = function(x)
               diff(sort(x))
           ))
-      },
+        belief_states
+      }, 
       
       regular = {
+        belief_states <-
+          matrix(0, nrow = n, ncol = length(model$states))
+        colnames(belief_states) <- model$states
+        
         if (d == 2) {
           b <- seq(0, 1, length.out = n)
           belief_states[, projection] <- cbind(b, 1 - b)
@@ -107,14 +129,28 @@ sample_belief_space <-
             triangleCentres
         } else
           stop("method redular is only available for projections on 2 or 3 states.")
+        
+        belief_states
       },
       
       vertices = {
+        belief_states <-
+          matrix(0, nrow = n, ncol = length(model$states))
+        colnames(belief_states) <- model$states
+        
         belief_states <- belief_states[rep(1, d),]
         belief_states[cbind(projection, projection)] <- 1
         belief_states <-
           belief_states[sample(d, size = n, replace = TRUE),]
-      })
+        
+        belief_states
+      },
+      
+      trajectories = {
+        simulate_POMDP(model, n = n, visited_beliefs = TRUE, ...) 
+      }
+      
+      )
     
     belief_states
   }